@@ -1,10 +1,26 @@
+//importeer het "fetch-json" bestand en geef het de naam fetchJson
 import fetchJson from "$lib/fetch-json"
 
-export async function load() {
-  const API = 'https://fdnd-agency.directus.app/items/'
-  const items = await fetchJson(API + 'deloitte_packages')
- 
+//een functie om data in te laden met async wacht je totdat the API request gedaan is
+export async function load({cookies}) {
+  //een variable met een APi url voor person met id 56
+  const questions_url = 'https://fdnd-agency.directus.app/items/deloitte_questions'
+  const answers_url = 'https://fdnd-agency.directus.app/items/deloitte_answers'
+  const packages_url = 'https://fdnd-agency.directus.app/items/deloitte_packages'
+  const users_url = 'https://fdnd-agency.directus.app/items/deloitte_users'
+  //fetch de data uit de API url en sla het op in deze variable
+  const questions = await fetchJson(questions_url)
+  const answers = await fetchJson(answers_url)
+  const packages = await fetchJson(packages_url)
+  const users = await fetchJson(users_url)
+  const userID = cookies.get('userID') || null;
+
+  //return een object waarin person.data wordt meegegeven
   return {
-    items: items.data
+    questions: questions.data,
+    answers: answers.data,
+    packages: packages.data,
+    users: users.data,
+    userID: userID
   }
 }
