@@ -1,6 +1,13 @@
 import { sveltekit } from '@sveltejs/kit/vite';
+import { enhancedImages } from '@sveltejs/enhanced-img';
 import { defineConfig } from 'vite';
 
 export default defineConfig({
-	plugins: [sveltekit()]
+	plugins: [
+		enhancedImages(),
+		sveltekit()
+	],
+	server: {
+		port: 7006,
+	},
 });
